feat: DB functional
We added minor functionality to DB from CRUD basic operations, we've just tried CREATE and DELETE. Also, we inflated the menu in the InventoryApp activity. At the end we started to work with the content providers in order to mix it with the cursorloader to display the information smoothly in a ListView.

Resolves: NA
See also: NA